mors/requirements.txt
cs50
Flask
Flask-Session
requests
datetime
SQL
qrcode
pillow
PyQRCode
pypng
cv2